GeneratorStressTest - stress test works! Runs the gauntlet.
On an HP EliteBook 8750w laptop with 32GB ram and default Eclipse Neon
Java configuration, it runs fine with 40 generators, each yielding 2000
times, and before each yield, each generator makes 400 busywork calls,
recursing to a depth of 200. Iterators are called in the correct
sequence and the client code is non-reentrant. So in other words, the
core generator library checks out well.