Add Thread.yield in iterator logic as well; change intFuncs; add a few more comments.
- Added a Thread.yield call to IntFuncGenerator's run method to help
precipitate threading errors if any.  The client code already had one.
- Switch intFuncs from generating multiples of successive primes to
simple (BASE*n + i) because the latter is arguably better for large
values of BASE and easier to programmitcally generate functions should
the need arise. Current value of BASE is 10,000, which means that for
less than 10,000 iterations all return values are unique.