First StructuredLogger JUnit tests pass!
It caught the fact I wasn't logging the session description in the endSession call. That was originally by design, but I decided to add the session description just to reinforce the session description on closing.

The first JUnit test - "testBeginEndSession" - which has no body because all the verification happens in the pre/post test code - works!

This is actually quite a bit of verification as begin/end session logs stuff and the logs are flushed and closed on endSession - all of this is verified.