FileLogger: revamped to take File objects and support two creation modes
The two creation modes are:
1. A single log file - that is specified by the user. With option to append.
2. One log file per session - created by the RawLogger. With option to append (though we would expect  a new file to be created each time because it is based on the session ID, which is getcurrenttimelillis.

Also, we write to System.err on any caught exceptions. We don't assert. If there is an exception caught in the log method we only write to System.err once so we don't clutter the err console.

JUnit tests have been created and pass for both types of FileLogger creation.