JUnit test for UDPRawLogger written and passes
This is testUDPRawLogger(). Sends three raw log messages and verifies that the messages are received as UDP datagrams. Uses the ExecutorService to receive a a background task. This may be unnecessary as the receive datagram may start receiving messages and buffering them before we actually attempt to receive them - need to verify this -- but it is also a quick trial of using the ExecutorService and ConcurrentLinkedQueue - which will get used in the implementation of the RobotComm class.