StructuredLogger: added err, warn info to the base StructuredLogger
This is for convenience - so simple logging requires the fewest steps.
Also: route internal calls to create a new log to a common method in the StructuredLogger (commonNewLog). This is so that in the future we can if we want do some global thing when create new logs like add them to a global list of logs. For the present, however, we do not keep track of created logs.