StructuredLogger: trigger flushing correctly; consolidate background processing tasks
Implemented logic for keeping track of un-flushed messages on a per-rawlog basis - this is the count of messages in the queue PLUS count of messages that havev been written but the log not flushed.

Consolidated various background TimerTask objects into a single method (newBackgroundProcessor) for creating a single type of background task.