StructuredLogger: Implemented and run logger stress test
- Fix typo in PRI1 (2->1)
Fixes/changes
- Keep track of total count of discarded messages for the logging session.
- rawLog(): add a space after "_msg:" so that subsequent kv pairs look nice
   Was: "_msg:a:b" now is "_msg: a:b". Note that both parse correctly, it's just cleaner
  to look at with the space there.

The major addition in this checkin is the StructuredLogger stress test "testStressTest". It currently sends 30000 messages from 3 threads. It can be configured to
force discarding of messages, but currently random sleeps have been enserted so
that the background logging tasks can keep up with the onslaught (at least on a Dell XPS i7 machine). When messages are discarded there is a small, relatively constant (10-14 regardless of number of messages sent as long as they are more than 10000) difference between the expected discarded and the actual discarded messages that is something to investigate further.