StructuredLogger: Stop logging on buffer limit reached  NOT trim old msgs, and other fixes.
We used to trim a certain fraction of  oldest messages if the logging limit was reached. We not simply don't log if this limit is reached. This seems a more intuitive way to deal with the situation and also will hopefully result in more fine-grained discarding of messages.

Earlier, LoggingImplementation.flush() was ALWAYS scheduling a one-shot background tasks. In the stress test, flush was sometimes being called 100s of times per second.

In the fix, all triggering of one-shot tasks are routed through triggerBackgroundTaskIfRunning, which doesn't schedule a task if there is one running (this logic was there previously - just in the rawLog() code).

Another fix was to log discarded messaged BEFORE we set sessionStarted to false (the log message about discarded messages was getting ignored if it happened on shutdown).

Changed TRIGGER_FRACTION from 0.5 to 0.25 - this is the fraction of buffer usage at which we trigger a background task to process the buffers.