RobotComm: added close method and commClosed field.
RobotComm.close() sets the commClosed field to false, and then goes through and closes all channels (closing a channel removes itself from the list of channels). So on exit to RobotComm.close() the channels list is empty. Attempts to add a new channel results in throwing an IllegalStateException exception.

There is a small window where newChannel is canned concurrently with RobotComm.close() - a channel can be created, but it will never be closed or removed from the list.  On the other hand, when the underlying transport is closed, underlying resources should be released.