RobotComm:: Channel.sendCommand: added addToCompletionQueue, other changes
- Added addToCompletionQueue to Channel.sendCommand, and added new method
Channel.pollCompletedCommand. This is because it will be convenient for client software to look for any completed sent command in one place, rather than check each
outstanding command for completion in an ad-hoc manner. However we give the client control of whether any particular instance of SentCommand is to be queued or not. Once it is queued, it MUST be removed from the queue or it will be in the queue indefinately.

IMPNOTES.MD: made extensive notes on handling of received commands - including changes to protocol. See February 14, 2018 design note.