RobotComm: Stress testing of commands - code completet
The code is in - see impnotes.md design note Feb 25A. The stress test doesn't run properly. It seems to run fine with submitting multiple commands with no delays or drops, but fails if random transport drops are included. Could be a problem with RobotComm or the stress testing code.