RobotComm: fix bug in randExpDelay
mixup of max and min. Now random exponential backoff seems too work - sent 1 million commands on a transport with 25% failure rate. There were about 200 missing commands that are expected to complete but have not by the end of the test, though I suspect that this is the test terminating too early. Needs to be investigated.