RobotComm: First implementation of server pruning completed commands
With this implementation (set at 1 million queued commands), the stress test sent 20 million commands over a transport with errors and delays and force-drops! See impnotes.md Feb 28B, 2018 note. Previously it would simply run out of memory because server would never prune completed commands. Still work to be done as 1-million was chosen to make the 20-million test pass. The fix is to implement CMDRESPACK, which is up next.