UdpTransport: keeps a concurrent hash map of remote nodes.
Priviously UdpTransport kept a HashMap of remote nodes that it created in demand  - this was used ONLY by the receive packet thread. In order to avail of this cash externally, I made it a concurrent hash map, and  newRemoteNode now accesses this cache. This method is called by RobotComm to implement sendMessage to an unbound remote note (privously that version of sendMessage was unimplemented).

Other changes: UdpTransport traces all sent/receive data.. Some cosmetic cleanup suggested by SonarLint.