rcping: change -payload option to -body and fixed default handling of set_parameters
EchoClient.set_parameters was incorrectly setting bodytype  default to 'hello'.  It
should be None so that if None is passed the right thing happens - which is
to change that to 'hello'.  rcping may pass in empty or None for bodytype and body.
The two are treated differently - empty bodytype results in an empty bodytype,
 while None results in the default bodytype which is 'hello'.